Revert to single-file source code.
