+++ main-cli.c
@@ -16,7 +16,6 @@
 #include <time.h>
 #include <popt.h>
 #include "mmio/mmio.h"
-#include "smvp-algs/smvp-algs.h"
 
 // ANSI terminal color escape codes for making output BEAUTIFUL
 #define ANSI_COLOR_RED "\x1b[31m"
@@ -40,6 +39,25 @@ typedef struct _mm_raw_data_
     int col;
     double val;
 } MMRawData;
+
+// Struct: _csr_data_
+// Provides a convenient structure for storing/manipulating CSR compressed data
+typedef struct _csr_data_
+{
+    int *row_ptr;
+    int *col_ind;
+    double *val;
+} CSRData;
+
+// Function: vectorInit
+// Reinitializes vectors between calculation iterations
+void vectorInit(int vectorLen, double *outputVector, double val)
+{
+    for (int index = 0; index < vectorLen; index++)
+    {
+        outputVector[index] = val;
+    }
+}
 
 // Function: mmioErrorHandler
 // Provides a simple error handler for known error types (mmio.h)
@@ -131,6 +149,94 @@ void generateReportFile()
 }
 */
 
+// Function: smvp_csr_compute
+// Calculates SMVP using CSR algorithm
+void smvp_csr_compute(MMRawData *mmImportData, int fInputRows, int fInputNonZeros)
+{
+
+    CSRData workingMatrix;
+    double *onesVector, *outputVector;
+    struct timespec tsCompStart, tsCompEnd;
+    double comp_time_taken;
+    int j, compIter, index;
+
+    workingMatrix.row_ptr = (int *)malloc(sizeof(int) * (long unsigned int)(fInputRows + 1));
+    workingMatrix.col_ind = (int *)malloc(sizeof(int) * (long unsigned int)fInputNonZeros);
+    workingMatrix.val = (double *)malloc(sizeof(double) * (long unsigned int)fInputNonZeros);
+
+    for (index = 0; index < fInputNonZeros; index++)
+    {
+        workingMatrix.val[index] = mmImportData[index].val;
+        workingMatrix.col_ind[index] = mmImportData[index].col;
+
+        if (index == fInputNonZeros - 1)
+        {
+            workingMatrix.row_ptr[mmImportData[index].row + 1] = fInputNonZeros;
+        }
+        else if (mmImportData[index].row < mmImportData[(index + 1)].row)
+        {
+            workingMatrix.row_ptr[mmImportData[index].row + 1] = index + 1;
+        }
+        else if (index == 0)
+        {
+            workingMatrix.row_ptr[mmImportData[index].row] = 0;
+        }
+    }
+
+    // Prepare the "ones "vector and output vector
+    onesVector = (double *)malloc(sizeof(double) * (long unsigned int)fInputRows);
+    vectorInit(fInputRows, onesVector, 1);
+    outputVector = (double *)malloc(sizeof(double) * (long unsigned int)fInputRows);
+
+    // Capture compute start time
+    clock_gettime(CLOCK_MONOTONIC, &tsCompStart);
+
+    // Compute the sparse vector multiplication (technically y=Axn, not y=x(A^n) as indicated in reqs doc, but is an approved deviation)
+    for (compIter = 0; compIter < 1000; compIter++)
+    {
+        //Reset output vector contents between iterations
+        vectorInit(fInputRows, outputVector, 0);
+
+        for (index = 0; index < fInputRows; index++)
+        {
+            for (j = workingMatrix.row_ptr[index]; j < workingMatrix.row_ptr[index + 1]; j++)
+            {
+                outputVector[index] += workingMatrix.val[j] * onesVector[workingMatrix.col_ind[j]];
+            }
+        }
+    }
+
+    // Capture compute end time & derive elapsed time
+    clock_gettime(CLOCK_MONOTONIC, &tsCompEnd);
+    comp_time_taken = (double)(tsCompEnd.tv_sec - tsCompStart.tv_sec) * 1e9;
+    comp_time_taken = (comp_time_taken + (double)(tsCompEnd.tv_nsec - tsCompStart.tv_nsec)) * 1e-9;
+
+    // RETURN SOMETHING
+}
+
+/*
+void smvp_csr_debug(FILE *reportOutputFile, int fInputRows)
+{
+    // Append debug info in output file if required
+    fprintf(reportOutputFile, "[DEBUG]\tCSR row_ptr:\n[");
+    for (index = 0; index < fInputRows + 1; index++)
+    {
+        fprintf(reportOutputFile, "%d\n", workingMatrix.row_ptr[index]);
+    }
+    fprintf(reportOutputFile, "]\n[DEBUG]\tCSR val:\n[");
+    for (index = 0; index < fInputNonZeros; index++)
+    {
+        fprintf(reportOutputFile, "%g\n", workingMatrix.val[index]);
+    }
+    fprintf(reportOutputFile, "]\n[DEBUG]\tCSR col_ind:\n[");
+    for (index = 0; index < fInputNonZeros; index++)
+    {
+        fprintf(reportOutputFile, "%d\n", workingMatrix.col_ind[index]);
+    }
+    fprintf(reportOutputFile, "]\n");
+}
+*/
+
 int main(int argc, const char *argv[])
 {
 
@@ -143,22 +249,20 @@ int main(int argc, const char *argv[])
     int *iteration_time;
     double *output_vector;
 
-    printf(ANSI_COLOR_GREEN "\n[START]\tExecuting smvp-toolbox-cli v%d.%d.%d\n" ANSI_COLOR_RESET, MAJOR_VER, MINOR_VER, REVISION_VER);
-
     // Ust POPT library to handle command line arguments robustly
     // POPT library and documentation available at https://github.com/devzero2000/POPT
     struct _popt_field
     {
         int iter;
         char *file;
-        /* etc */
+
     } popt_field;
 
     struct poptOption optionsTable[] = {
         {"all", 'a', POPT_ARG_NONE, NULL, 'a', "Enable all SMVP algorithms."},
         {"csr", 'c', POPT_ARG_NONE, NULL, 'c', "Enable CSR SMVP algorithm."},
         {"tjds", 't', POPT_ARG_NONE, NULL, 't', "Enable TJDS SMVP algorithm."},
-        {"iter", 'i', POPT_ARG_INT, &popt_field.iter, 'i', "Specify computation iterations."},
+        {"iter", 'i', POPT_ARG_INT | POPT_ARGFLAG_OPTIONAL, &popt_field.iter, 'i', "Specify computation iterations."},
         {"file", 'f', POPT_ARG_STRING, &popt_field.file, 'f', "Specify path to input matrix file."},
         POPT_AUTOHELP
             POPT_TABLEEND};
@@ -174,14 +278,38 @@ int main(int argc, const char *argv[])
         switch (c)
         {
         case 'a':
-            alg_mode = ALG_ALL;
-            break;
+            if (alg_mode != ALG_NONE)
+            {
+                printf(ANSI_COLOR_RED "[ERROR]\tCombining [-a|--all] with other algorithm flags is not supported." ANSI_COLOR_RESET);
+                exit(1);
+            }
+            else
+            {
+                alg_mode = ALG_ALL;
+                break;
+            }
         case 'c':
-            alg_mode += ALG_CSR;
-            break;
+            if (alg_mode == ALG_ALL)
+            {
+                printf(ANSI_COLOR_RED "[ERROR]\tCombining [-a|--all] with other algorithm flags is not supported." ANSI_COLOR_RESET);
+                exit(1);
+            }
+            else
+            {
+                alg_mode += ALG_CSR;
+                break;
+            }
         case 't':
-            alg_mode += ALG_TJDS;
-            break;
+            if (alg_mode == ALG_ALL)
+            {
+                printf(ANSI_COLOR_RED "[ERROR]\tCombining [-a|--all] with other algorithm flags is not supported." ANSI_COLOR_RESET);
+                exit(1);
+            }
+            else
+            {
+                alg_mode += ALG_TJDS;
+                break;
+            }
         case 'i':
             calc_iter = popt_field.iter;
             break;
@@ -201,6 +329,8 @@ int main(int argc, const char *argv[])
             exit(1);
         }
     }
+
+    printf(ANSI_COLOR_GREEN "\n[START]\tExecuting smvp-toolbox-cli v%d.%d.%d\n" ANSI_COLOR_RESET, MAJOR_VER, MINOR_VER, REVISION_VER);
 
     // Ensure input file is of proper format and contains an appropriate matrix type
     mmio_rb_return = mm_read_banner(mmInputFile, &matcode);
